split build.ts, add comments, fix lint
