+++ @types/bun/index.d.ts
@@ -1,0 +1,2 @@
+declare module 'bun:ffi'
+declare module 'bun:test'
