+++ shared-native/build_helper/const.ts
@@ -1,0 +1,1 @@
+export const ZIG_VERSION = '0.10.1'
